sync master branch

+++ test/e2e_node/cpu_manager_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package e2e_node
+package e2enode
 
 import (
 	"fmt"
@@ -23,18 +23,21 @@ import (
 	"strings"
 	"time"
 
-	"k8s.io/api/core/v1"
+	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	runtimeapi "k8s.io/kubernetes/pkg/kubelet/apis/cri/runtime/v1alpha2"
-	"k8s.io/kubernetes/pkg/kubelet/apis/kubeletconfig"
+	runtimeapi "k8s.io/cri-api/pkg/apis/runtime/v1alpha2"
+	kubeletconfig "k8s.io/kubernetes/pkg/kubelet/apis/config"
 	"k8s.io/kubernetes/pkg/kubelet/cm/cpumanager"
+	cpumanagerstate "k8s.io/kubernetes/pkg/kubelet/cm/cpumanager/state"
 	"k8s.io/kubernetes/pkg/kubelet/cm/cpuset"
 	"k8s.io/kubernetes/pkg/kubelet/types"
 	"k8s.io/kubernetes/test/e2e/framework"
+	e2enode "k8s.io/kubernetes/test/e2e/framework/node"
+	e2eskipper "k8s.io/kubernetes/test/e2e/framework/skipper"
 
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
+	"github.com/onsi/ginkgo"
+	"github.com/onsi/gomega"
 )
 
 // Helper for makeCPUManagerPod().
@@ -93,7 +96,7 @@ func getLocalNodeCPUDetails(f *framework.Framework) (cpuCapVal int64, cpuAllocVa
 	cpuCap := localNodeCap[v1.ResourceCPU]
 	localNodeAlloc := getLocalNode(f).Status.Allocatable
 	cpuAlloc := localNodeAlloc[v1.ResourceCPU]
-	cpuRes := cpuCap.Copy()
+	cpuRes := cpuCap.DeepCopy()
 	cpuRes.Sub(cpuAlloc)
 
 	// RoundUp reserved CPUs to get only integer cores.
@@ -104,8 +107,8 @@ func getLocalNodeCPUDetails(f *framework.Framework) (cpuCapVal int64, cpuAllocVa
 
 func waitForContainerRemoval(containerName, podName, podNS string) {
 	rs, _, err := getCRIClient()
-	Expect(err).NotTo(HaveOccurred())
-	Eventually(func() bool {
+	framework.ExpectNoError(err)
+	gomega.Eventually(func() bool {
 		containers, err := rs.ListContainers(&runtimeapi.ContainerFilter{
 			LabelSelector: map[string]string{
 				types.KubernetesPodNameLabel:       podName,
@@ -117,7 +120,19 @@ func waitForContainerRemoval(containerName, podName, podNS string) {
 			return false
 		}
 		return len(containers) == 0
-	}, 2*time.Minute, 1*time.Second).Should(BeTrue())
+	}, 2*time.Minute, 1*time.Second).Should(gomega.BeTrue())
+}
+
+func waitForStateFileCleanedUp() {
+	gomega.Eventually(func() bool {
+		restoredState, err := cpumanagerstate.NewCheckpointState("/var/lib/kubelet", "cpu_manager_state", "static", nil)
+		framework.ExpectNoError(err, "failed to create testing cpumanager state instance")
+		assignments := restoredState.GetCPUAssignments()
+		if len(assignments) == 0 {
+			return true
+		}
+		return false
+	}, 2*time.Minute, 1*time.Second).Should(gomega.BeTrue())
 }
 
 func isHTEnabled() bool {
@@ -151,13 +166,38 @@ func setOldKubeletConfig(f *framework.Framework, oldCfg *kubeletconfig.KubeletCo
 	}
 }
 
-func enableCPUManagerInKubelet(f *framework.Framework) (oldCfg *kubeletconfig.KubeletConfiguration) {
+func disableCPUManagerInKubelet(f *framework.Framework) (oldCfg *kubeletconfig.KubeletConfiguration) {
+	// Disable CPU Manager in Kubelet.
+	oldCfg, err := getCurrentKubeletConfig()
+	framework.ExpectNoError(err)
+	newCfg := oldCfg.DeepCopy()
+	if newCfg.FeatureGates == nil {
+		newCfg.FeatureGates = make(map[string]bool)
+	}
+	newCfg.FeatureGates["CPUManager"] = false
+
+	// Update the Kubelet configuration.
+	framework.ExpectNoError(setKubeletConfiguration(f, newCfg))
+
+	// Wait for the Kubelet to be ready.
+	gomega.Eventually(func() bool {
+		nodes, err := e2enode.TotalReady(f.ClientSet)
+		framework.ExpectNoError(err)
+		return nodes == 1
+	}, time.Minute, time.Second).Should(gomega.BeTrue())
+
+	return oldCfg
+}
+
+func enableCPUManagerInKubelet(f *framework.Framework, cleanStateFile bool) (oldCfg *kubeletconfig.KubeletConfiguration) {
 	// Enable CPU Manager in Kubelet with static policy.
 	oldCfg, err := getCurrentKubeletConfig()
 	framework.ExpectNoError(err)
 	newCfg := oldCfg.DeepCopy()
 	if newCfg.FeatureGates == nil {
 		newCfg.FeatureGates = make(map[string]bool)
+	} else {
+		newCfg.FeatureGates["CPUManager"] = true
 	}
 
 	// After graduation of the CPU Manager feature to Beta, the CPU Manager
@@ -168,7 +208,10 @@ func enableCPUManagerInKubelet(f *framework.Framework) (oldCfg *kubeletconfig.Ku
 	// be "none" whereas we are trying to restart Kubelet with "static"
 	// policy). Therefore, we delete the state file so that we can proceed
 	// with the tests.
-	deleteStateFile()
+	// Only delete the state file at the begin of the tests.
+	if cleanStateFile {
+		deleteStateFile()
+	}
 
 	// Set the CPU Manager policy to static.
 	newCfg.CPUManagerPolicy = string(cpumanager.PolicyStatic)
@@ -190,10 +233,11 @@ func enableCPUManagerInKubelet(f *framework.Framework) (oldCfg *kubeletconfig.Ku
 	framework.ExpectNoError(setKubeletConfiguration(f, newCfg))
 
 	// Wait for the Kubelet to be ready.
-	Eventually(func() bool {
-		nodeList := framework.GetReadySchedulableNodesOrDie(f.ClientSet)
-		return len(nodeList.Items) == 1
-	}, time.Minute, time.Second).Should(BeTrue())
+	gomega.Eventually(func() bool {
+		nodes, err := e2enode.TotalReady(f.ClientSet)
+		framework.ExpectNoError(err)
+		return nodes == 1
+	}, time.Minute, time.Second).Should(gomega.BeTrue())
 
 	return oldCfg
 }
@@ -209,18 +253,18 @@ func runCPUManagerTests(f *framework.Framework) {
 	var ctnAttrs []ctnAttribute
 	var pod, pod1, pod2 *v1.Pod
 
-	It("should assign CPUs as expected based on the Pod spec", func() {
+	ginkgo.It("should assign CPUs as expected based on the Pod spec", func() {
 		cpuCap, cpuAlloc, _ = getLocalNodeCPUDetails(f)
 
 		// Skip CPU Manager tests altogether if the CPU capacity < 2.
 		if cpuCap < 2 {
-			framework.Skipf("Skipping CPU Manager tests since the CPU capacity < 2")
+			e2eskipper.Skipf("Skipping CPU Manager tests since the CPU capacity < 2")
 		}
 
 		// Enable CPU Manager in the kubelet.
-		oldCfg = enableCPUManagerInKubelet(f)
+		oldCfg = enableCPUManagerInKubelet(f, true)
 
-		By("running a non-Gu pod")
+		ginkgo.By("running a non-Gu pod")
 		ctnAttrs = []ctnAttribute{
 			{
 				ctnName:    "non-gu-container",
@@ -231,17 +275,17 @@ func runCPUManagerTests(f *framework.Framework) {
 		pod = makeCPUManagerPod("non-gu-pod", ctnAttrs)
 		pod = f.PodClient().CreateSync(pod)
 
-		By("checking if the expected cpuset was assigned")
+		ginkgo.By("checking if the expected cpuset was assigned")
 		expAllowedCPUsListRegex = fmt.Sprintf("^0-%d\n$", cpuCap-1)
 		err = f.PodClient().MatchContainerOutput(pod.Name, pod.Spec.Containers[0].Name, expAllowedCPUsListRegex)
 		framework.ExpectNoError(err, "expected log not found in container [%s] of pod [%s]",
 			pod.Spec.Containers[0].Name, pod.Name)
 
-		By("by deleting the pods and waiting for container removal")
+		ginkgo.By("by deleting the pods and waiting for container removal")
 		deletePods(f, []string{pod.Name})
 		waitForContainerRemoval(pod.Spec.Containers[0].Name, pod.Name, pod.Namespace)
 
-		By("running a Gu pod")
+		ginkgo.By("running a Gu pod")
 		ctnAttrs = []ctnAttribute{
 			{
 				ctnName:    "gu-container",
@@ -252,7 +296,7 @@ func runCPUManagerTests(f *framework.Framework) {
 		pod = makeCPUManagerPod("gu-pod", ctnAttrs)
 		pod = f.PodClient().CreateSync(pod)
 
-		By("checking if the expected cpuset was assigned")
+		ginkgo.By("checking if the expected cpuset was assigned")
 		cpu1 = 1
 		if isHTEnabled() {
 			cpuList = cpuset.MustParse(getCPUSiblingList(0)).ToSlice()
@@ -263,11 +307,11 @@ func runCPUManagerTests(f *framework.Framework) {
 		framework.ExpectNoError(err, "expected log not found in container [%s] of pod [%s]",
 			pod.Spec.Containers[0].Name, pod.Name)
 
-		By("by deleting the pods and waiting for container removal")
+		ginkgo.By("by deleting the pods and waiting for container removal")
 		deletePods(f, []string{pod.Name})
 		waitForContainerRemoval(pod.Spec.Containers[0].Name, pod.Name, pod.Namespace)
 
-		By("running multiple Gu and non-Gu pods")
+		ginkgo.By("running multiple Gu and non-Gu pods")
 		ctnAttrs = []ctnAttribute{
 			{
 				ctnName:    "gu-container",
@@ -288,7 +332,7 @@ func runCPUManagerTests(f *framework.Framework) {
 		pod2 = makeCPUManagerPod("non-gu-pod", ctnAttrs)
 		pod2 = f.PodClient().CreateSync(pod2)
 
-		By("checking if the expected cpuset was assigned")
+		ginkgo.By("checking if the expected cpuset was assigned")
 		cpu1 = 1
 		if isHTEnabled() {
 			cpuList = cpuset.MustParse(getCPUSiblingList(0)).ToSlice()
@@ -309,17 +353,17 @@ func runCPUManagerTests(f *framework.Framework) {
 		framework.ExpectNoError(err, "expected log not found in container [%s] of pod [%s]",
 			pod2.Spec.Containers[0].Name, pod2.Name)
 
-		By("by deleting the pods and waiting for container removal")
+		ginkgo.By("by deleting the pods and waiting for container removal")
 		deletePods(f, []string{pod1.Name, pod2.Name})
 		waitForContainerRemoval(pod1.Spec.Containers[0].Name, pod1.Name, pod1.Namespace)
 		waitForContainerRemoval(pod2.Spec.Containers[0].Name, pod2.Name, pod2.Namespace)
 
 		// Skip rest of the tests if CPU capacity < 3.
 		if cpuCap < 3 {
-			framework.Skipf("Skipping rest of the CPU Manager tests since CPU capacity < 3")
+			e2eskipper.Skipf("Skipping rest of the CPU Manager tests since CPU capacity < 3")
 		}
 
-		By("running a Gu pod requesting multiple CPUs")
+		ginkgo.By("running a Gu pod requesting multiple CPUs")
 		ctnAttrs = []ctnAttribute{
 			{
 				ctnName:    "gu-container",
@@ -330,7 +374,7 @@ func runCPUManagerTests(f *framework.Framework) {
 		pod = makeCPUManagerPod("gu-pod", ctnAttrs)
 		pod = f.PodClient().CreateSync(pod)
 
-		By("checking if the expected cpuset was assigned")
+		ginkgo.By("checking if the expected cpuset was assigned")
 		cpuListString = "1-2"
 		if isHTEnabled() {
 			cpuListString = "2-3"
@@ -345,11 +389,11 @@ func runCPUManagerTests(f *framework.Framework) {
 		framework.ExpectNoError(err, "expected log not found in container [%s] of pod [%s]",
 			pod.Spec.Containers[0].Name, pod.Name)
 
-		By("by deleting the pods and waiting for container removal")
+		ginkgo.By("by deleting the pods and waiting for container removal")
 		deletePods(f, []string{pod.Name})
 		waitForContainerRemoval(pod.Spec.Containers[0].Name, pod.Name, pod.Namespace)
 
-		By("running a Gu pod with multiple containers requesting integer CPUs")
+		ginkgo.By("running a Gu pod with multiple containers requesting integer CPUs")
 		ctnAttrs = []ctnAttribute{
 			{
 				ctnName:    "gu-container1",
@@ -365,7 +409,7 @@ func runCPUManagerTests(f *framework.Framework) {
 		pod = makeCPUManagerPod("gu-pod", ctnAttrs)
 		pod = f.PodClient().CreateSync(pod)
 
-		By("checking if the expected cpuset was assigned")
+		ginkgo.By("checking if the expected cpuset was assigned")
 		cpu1, cpu2 = 1, 2
 		if isHTEnabled() {
 			cpuList = cpuset.MustParse(getCPUSiblingList(0)).ToSlice()
@@ -379,16 +423,16 @@ func runCPUManagerTests(f *framework.Framework) {
 		framework.ExpectNoError(err, "expected log not found in container [%s] of pod [%s]",
 			pod.Spec.Containers[0].Name, pod.Name)
 
-		err = f.PodClient().MatchContainerOutput(pod.Name, pod.Spec.Containers[0].Name, expAllowedCPUsListRegex)
+		err = f.PodClient().MatchContainerOutput(pod.Name, pod.Spec.Containers[1].Name, expAllowedCPUsListRegex)
 		framework.ExpectNoError(err, "expected log not found in container [%s] of pod [%s]",
 			pod.Spec.Containers[1].Name, pod.Name)
 
-		By("by deleting the pods and waiting for container removal")
+		ginkgo.By("by deleting the pods and waiting for container removal")
 		deletePods(f, []string{pod.Name})
 		waitForContainerRemoval(pod.Spec.Containers[0].Name, pod.Name, pod.Namespace)
 		waitForContainerRemoval(pod.Spec.Containers[1].Name, pod.Name, pod.Namespace)
 
-		By("running multiple Gu pods")
+		ginkgo.By("running multiple Gu pods")
 		ctnAttrs = []ctnAttribute{
 			{
 				ctnName:    "gu-container1",
@@ -409,7 +453,7 @@ func runCPUManagerTests(f *framework.Framework) {
 		pod2 = makeCPUManagerPod("gu-pod2", ctnAttrs)
 		pod2 = f.PodClient().CreateSync(pod2)
 
-		By("checking if the expected cpuset was assigned")
+		ginkgo.By("checking if the expected cpuset was assigned")
 		cpu1, cpu2 = 1, 2
 		if isHTEnabled() {
 			cpuList = cpuset.MustParse(getCPUSiblingList(0)).ToSlice()
@@ -428,10 +472,53 @@ func runCPUManagerTests(f *framework.Framework) {
 		framework.ExpectNoError(err, "expected log not found in container [%s] of pod [%s]",
 			pod2.Spec.Containers[0].Name, pod2.Name)
 
-		By("by deleting the pods and waiting for container removal")
+		ginkgo.By("by deleting the pods and waiting for container removal")
 		deletePods(f, []string{pod1.Name, pod2.Name})
 		waitForContainerRemoval(pod1.Spec.Containers[0].Name, pod1.Name, pod1.Namespace)
 		waitForContainerRemoval(pod2.Spec.Containers[0].Name, pod2.Name, pod2.Namespace)
+
+		ginkgo.By("test for automatically remove inactive pods from cpumanager state file.")
+		// First running a Gu Pod,
+		// second disable cpu manager in kubelet,
+		// then delete the Gu Pod,
+		// then enable cpu manager in kubelet,
+		// at last wait for the reconcile process cleaned up the state file, if the assignments map is empty,
+		// it proves that the automatic cleanup in the reconcile process is in effect.
+		ginkgo.By("running a Gu pod for test remove")
+		ctnAttrs = []ctnAttribute{
+			{
+				ctnName:    "gu-container-testremove",
+				cpuRequest: "1000m",
+				cpuLimit:   "1000m",
+			},
+		}
+		pod = makeCPUManagerPod("gu-pod-testremove", ctnAttrs)
+		pod = f.PodClient().CreateSync(pod)
+
+		ginkgo.By("checking if the expected cpuset was assigned")
+		cpu1 = 1
+		if isHTEnabled() {
+			cpuList = cpuset.MustParse(getCPUSiblingList(0)).ToSlice()
+			cpu1 = cpuList[1]
+		}
+		expAllowedCPUsListRegex = fmt.Sprintf("^%d\n$", cpu1)
+		err = f.PodClient().MatchContainerOutput(pod.Name, pod.Spec.Containers[0].Name, expAllowedCPUsListRegex)
+		framework.ExpectNoError(err, "expected log not found in container [%s] of pod [%s]",
+			pod.Spec.Containers[0].Name, pod.Name)
+
+		ginkgo.By("disable cpu manager in kubelet")
+		disableCPUManagerInKubelet(f)
+
+		ginkgo.By("by deleting the pod and waiting for container removal")
+		deletePods(f, []string{pod.Name})
+		waitForContainerRemoval(pod.Spec.Containers[0].Name, pod.Name, pod.Namespace)
+
+		ginkgo.By("enable cpu manager in kubelet without delete state file")
+		enableCPUManagerInKubelet(f, false)
+
+		ginkgo.By("wait for the deleted pod to be cleaned up from the state file")
+		waitForStateFileCleanedUp()
+		ginkgo.By("the deleted pod has already been deleted from the state file")
 
 		setOldKubeletConfig(f, oldCfg)
 	})
@@ -441,7 +528,7 @@ func runCPUManagerTests(f *framework.Framework) {
 var _ = SIGDescribe("CPU Manager [Serial] [Feature:CPUManager][NodeAlphaFeature:CPUManager]", func() {
 	f := framework.NewDefaultFramework("cpu-manager-test")
 
-	Context("With kubeconfig updated with static CPU Manager policy run the CPU Manager tests", func() {
+	ginkgo.Context("With kubeconfig updated with static CPU Manager policy run the CPU Manager tests", func() {
 		runCPUManagerTests(f)
 	})
 })
